requirements.txt
africastalking==1.2.3
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
contextlib2==0.6.0.post1
cryptography==3.4.6
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.1.7
django-braces==1.14.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.4.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
schema==0.7.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
